Conflicts:
	server/factory/models/Boards.php

+++ client/app/routes/login_route.js
@@ -1,6 +1,14 @@
 App.LoginRoute = Ember.Route.extend({
 	renderTemplate:function() {
 		this._super();
-		localStorage.clear();
+		if(!localStorage['reloaded']) {
+			localStorage.clear();
+			localStorage['reloaded'] = 'true';
+			Ember.run.scheduleOnce('afterRender', this, function() {
+				window.location.reload();
+			});
+		} else {
+			delete localStorage['reloaded'];
+		}
 	}
 });
